requirements.txt
# Roommate Agreement Generator - Dependencies
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database (MySQL)
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
cryptography>=41.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
httpx>=0.26.0

# Azure Services
azure-storage-blob>=12.19.0
azure-communication-email>=1.0.0

# Payments
stripe>=7.10.0
requests>=2.31.0

# DocuSign
docusign-esign>=3.25.0

# PDF Generation (optional)
jinja2>=3.1.0
# weasyprint>=60.0  # Requires system dependencies

# Development
python-multipart>=0.0.6
